+++ CHANGELOG.md
@@ -5,6 +5,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Create a new class for Layer 7 of OSI Model.
+- Add banner in `README.md` file
+- Create a new exception, `UninterestingPacketException`
+- Add command-line argument for specifying the desired network interface,
+`-i/--interface`
+- Create `decorators.py` file inside `utils` package that will contain, you've
+guessed it: decorators. Also, created the first decorator, `require_root` that
+will exit if the user running the script isn't **root**
+- Implement `-l/--list-interfaces` option, which will list all interfaces
+present in the system together with their MAC address + IPv4/6 address
+
+### Changed
+- Update Sniffer's description in `README.md` file
+
+### Fixed
+- Fix Layer 4 header size being (wrongly) too broad
 
 ## [0.2.0] - 2021-12-19
 ### Added